Update validation in make_move
Made make_move raise errors in more situations than
return the game with a message about error. When someone
guesses a letter already in the word, leaving that the same
since they guessed what would be a valid guess except they
forgot they already guessed it and need a reminder.

Also renamed variable in cancel_game to be more consistent.